requirements.txt
chainlit
openai
python-dotenv
requests
langchain
langchain-openai
langchain-community
langchain-mcp-adapters
langchain-tavily
beautifulsoup4
reportlab
openpyxl
python-docx
pypdf
boto3
asyncpg
asyncpg
SQLAlchemy
aiohttp 
greenlet